+++ shuup_tests/admin/test_product_mass_edit.py
@@ -31,7 +31,9 @@ def test_mass_edit_products(rf, admin_user):
     assert shop_product1.primary_category is None
     assert shop_product2.primary_category is None
 
-    request = apply_request_middleware(rf.post("/", data={"primary_category": category.pk}), user=admin_user)
+    request = apply_request_middleware(
+        rf.post("/", data={"primary_category": category.pk, "categories": [category.pk]}), user=admin_user
+    )
     request.session["mass_action_ids"] = [shop_product1.pk, shop_product2.pk]
 
     view = ProductMassEditView.as_view()
@@ -39,6 +41,7 @@ def test_mass_edit_products(rf, admin_user):
     assert response.status_code == 302
     for product in Product.objects.all():
         assert product.get_shop_instance(shop).primary_category == category
+        assert product.get_shop_instance(shop).categories.first() == category
 
 
 @pytest.mark.django_db